Refactoring ListenForUpdates to Use coroutine Channels
After refactoring our repository to use a ConflatedBroadcastChannel in place of our listener pub sub interfaces we need to update our command. Here we can simply subscribe to the repository before transforming the data as before but return a transformed ReceiveChannel. Clients can then use the pub sub API of the receive channel to consume repository events transformed in the command with our business rules.

Because of this we can also delete our EventEmitter interface and implementations. We do need to keep the Event.map extension though, so I moved it into the Extensions file in the Event package. keeping it as an extension allows us to still cover it with unit test coverage while excluding the class itself - which would be hard to cover because of the equals and hashcode methods needed for the immutable data structure nature of the class
