+++ domain/src/main/java/com/github/ianellis/shifts/domain/ListenForShiftUpdatesCommand.kt
@@ -1,10 +1,11 @@
 package com.github.ianellis.shifts.domain
 
 import com.github.ianellis.shifts.domain.events.Event
-import com.github.ianellis.shifts.domain.events.Ignore
 import com.github.ianellis.shifts.domain.events.map
 import com.github.ianellis.shifts.domain.time.Clock
 import com.github.ianellis.shifts.entities.ShiftEntity
+import kotlinx.coroutines.channels.ReceiveChannel
+import kotlinx.coroutines.channels.map
 import java.net.URL
 import java.text.SimpleDateFormat
 import java.util.*
@@ -12,21 +13,19 @@ import java.util.*
 class ListenForShiftUpdatesCommand constructor(
     private val shiftRepository: ShiftRepository,
     private val clock: Clock
-) : ((Event<List<Shift>>) -> Unit) -> Ignore {
+) : () -> ReceiveChannel<Event<List<Shift>>> {
 
     companion object {
         private const val ISO8601Format = "yyyy-MM-dd'T'HH:mm:ssZ"
     }
 
-    override fun invoke(listener: (Event<List<Shift>>) -> Unit): Ignore {
-        val repoListener: (Event<List<ShiftEntity>>) -> Unit = { event ->
-            val shifts = event.map { entityList ->
+    override fun invoke(): ReceiveChannel<Event<List<Shift>>> {
+        return shiftRepository.broadcastChannel.openSubscription().map { event ->
+            event.map { entityList ->
                 entityList.map { entity -> entity.toShift() }
             }
-            listener(shifts)
         }
-        shiftRepository.addListener(repoListener)
-        return { shiftRepository.removeListener(repoListener) }
+
     }
 
     private fun ShiftEntity.toShift(): Shift {
@@ -51,7 +50,7 @@ class ListenForShiftUpdatesCommand constructor(
         return format.parse(this.makeAndroidDateTimeCompliant())
     }
 
-    private fun String.makeAndroidDateTimeCompliant():String{
+    private fun String.makeAndroidDateTimeCompliant(): String {
         return this.replaceFirst(Regex("(\\d\\d):(\\d\\d)$"), "$1$2")
     }
 }
