+++ domain/src/test/java/com/github/ianellis/shifts/domain/ListenForShiftUpdatesCommandSpec.kt
@@ -8,7 +8,12 @@ import io.mockk.every
 import io.mockk.mockk
 import io.mockk.slot
 import io.mockk.verify
+import kotlinx.coroutines.channels.BroadcastChannel
+import kotlinx.coroutines.channels.ConflatedBroadcastChannel
+import kotlinx.coroutines.channels.consumeEach
+import kotlinx.coroutines.runBlocking
 import org.amshove.kluent.shouldBe
+import org.amshove.kluent.shouldEqual
 import org.junit.Before
 import org.junit.Test
 import java.net.URL
@@ -17,6 +22,7 @@ import java.util.*
 class ListenForShiftUpdatesCommandSpec {
 
     private lateinit var repository: ShiftRepository
+    private lateinit var broadcastChannel: ConflatedBroadcastChannel<Event<List<ShiftEntity>>>
     private lateinit var clock: Clock
 
     private lateinit var command: ListenForShiftUpdatesCommand
@@ -28,6 +34,10 @@ class ListenForShiftUpdatesCommandSpec {
     @Before
     fun setup() {
         repository = mockk(relaxUnitFun = true)
+
+        broadcastChannel = ConflatedBroadcastChannel()
+        every { repository.broadcastChannel } returns broadcastChannel
+
         clock = object : Clock {
             override fun now() = Date(TEST_TIME)
             override fun timezone() = TimeZone.getTimeZone("Australia/Sydney")
@@ -38,18 +48,7 @@ class ListenForShiftUpdatesCommandSpec {
 
     @Test
     fun `invoke() registers listener for events and maps repository data to Shift Entity`() {
-        //given
-        val repositoryListenerSlot = slot<(Event<List<ShiftEntity>>) -> Unit>()
-        every { repository.addListener(capture(repositoryListenerSlot)) } returns Unit
-
-        //when we invoke the command
-        val listener = mockFunction<Event<List<Shift>>>()
-        command(listener)
-
-        //then we register a listener in the repository
-        repositoryListenerSlot.isCaptured shouldBe true
-
-        // when the repository listener receives an event
+        // given the following entities
         val imageUrl1 = "http://google.com/some1.jpg"
         val imageUrl2 = "http://google.com/some2.jpg"
 
@@ -57,7 +56,11 @@ class ListenForShiftUpdatesCommandSpec {
             ShiftEntity(2, "2019-04-02T08:15:43+11:00", "", "0.00", "0.00", "", "", imageUrl2),
             ShiftEntity(1, "2019-03-31T21:15:18+11:00", "2019-03-31T22:15:43+11:00", "0.00", "0.00", "0.00", "0.00", imageUrl1)
         )
-        repositoryListenerSlot.captured(Event(shiftEntities))
+
+        //when the repository broadcasts an event
+        runBlocking {
+            broadcastChannel.send(Event(shiftEntities))
+        }
 
         //then we receive mapped Shift data objects
         val expectedShifts = listOf(
@@ -65,25 +68,9 @@ class ListenForShiftUpdatesCommandSpec {
             Shift(1, ShiftState.Complete(Date(1554027318000), Date(1554030943000)), URL(imageUrl1))
         )
 
-        verify { listener(Event(expectedShifts)) }
-    }
+        runBlocking {
+            command().receive() shouldEqual Event(expectedShifts)
+        }
 
-    @Test
-    fun `invoke() returns function to ignore events`() {
-        //given
-        val repositoryListenerSlot = slot<(Event<List<ShiftEntity>>) -> Unit>()
-        every { repository.addListener(capture(repositoryListenerSlot)) } returns Unit
-
-        //when we invoke the command
-        val ignore = command(mockFunction())
-
-        //then we register a listener in the repository
-        repositoryListenerSlot.isCaptured shouldBe true
-
-        // when we invoke the returned ignore function
-        ignore()
-
-        //then we unregister the listener from the repository
-        repository.removeListener(repositoryListenerSlot.captured)
     }
 }
